Add full config support to server.
