--- server/.env.docker
@@ -1,15 +1,0 @@
-# Environment variables for Docker Compose setup
-# Copy this file to .env and fill in your OpenAI API key
-
-# Required: OpenAI API Key for AI functionality
-OPENAI_API_KEY=your_openai_api_key_here
-
-# Optional: Browser type - will use default browser provided by Browserless
-# BROWSER=chromium
-
-# Optional: Override default port for Spark server
-# PORT=3000
-
-# Optional: Override browserless configuration
-# MAX_CONCURRENT_SESSIONS=5
-# CONNECTION_TIMEOUT=60000
