+++ server/.env.example
@@ -1,5 +1,143 @@
-# Server configuration
+# =============================================================================
+# SPARK SERVER CONFIGURATION
+# =============================================================================
+# Copy this file to .env and configure for your environment
+# 
+# Configuration Priority: Request parameters > Environment variables > Global config
+
+# =============================================================================
+# REQUIRED CONFIGURATION
+# =============================================================================
+
+# AI Provider API Key (choose one)
+OPENAI_API_KEY=your_openai_api_key_here
+# OPENROUTER_API_KEY=your_openrouter_api_key_here
+
+# =============================================================================
+# SERVER CONFIGURATION
+# =============================================================================
+
+# Server port
 PORT=3000
 
-# Spark configuration
-OPENAI_API_KEY=your_openai_api_key_here
+# =============================================================================
+# AI PROVIDER CONFIGURATION
+# =============================================================================
+
+# AI provider selection (default: openai)
+SPARK_PROVIDER=openai
+# SPARK_PROVIDER=openrouter
+
+# Model selection (uses provider defaults if not specified)
+# SPARK_MODEL=gpt-4.1                    # OpenAI
+# SPARK_MODEL=anthropic/claude-3-sonnet  # OpenRouter
+
+# =============================================================================
+# BROWSER CONFIGURATION
+# =============================================================================
+
+# Browser type (default: firefox)
+# Options: firefox, chrome, chromium, safari, webkit, edge
+SPARK_BROWSER=firefox
+
+# Headless mode (default: false for dev, true for production)
+SPARK_HEADLESS=false
+# SPARK_HEADLESS=true  # Recommended for Docker/production
+
+# Ad blocking (default: false)
+SPARK_BLOCK_ADS=true
+
+# Block resource types for better performance (comma-separated)
+# Options: image, stylesheet, font, media, manifest
+SPARK_BLOCK_RESOURCES=media,manifest
+
+# =============================================================================
+# WEBAGENT BEHAVIOR
+# =============================================================================
+
+# Debug mode with detailed logging (default: false)
+SPARK_DEBUG=false
+
+# Vision mode for screenshot analysis (default: false)
+SPARK_VISION=false
+
+# Maximum iterations (default: 50, consider 20-30 for containers)
+SPARK_MAX_ITERATIONS=50
+# SPARK_MAX_ITERATIONS=30  # For Docker/production
+
+# Maximum validation attempts (default: 3, consider 2 for containers)
+SPARK_MAX_VALIDATION_ATTEMPTS=3
+# SPARK_MAX_VALIDATION_ATTEMPTS=2  # For Docker/production
+
+# =============================================================================
+# PLAYWRIGHT CONFIGURATION
+# =============================================================================
+
+# Remote browser endpoint (optional)
+# SPARK_PW_ENDPOINT=ws://localhost:9222
+# SPARK_PW_ENDPOINT=ws://browserless:3000  # For Docker Compose
+
+# Bypass Content Security Policy (default: true)
+SPARK_BYPASS_CSP=true
+
+# =============================================================================
+# PROXY CONFIGURATION (Optional)
+# =============================================================================
+
+# Proxy server (supports http, https, socks5)
+# SPARK_PROXY=http://proxy.company.com:8080
+# SPARK_PROXY=socks5://127.0.0.1:1080
+
+# Proxy authentication
+# SPARK_PROXY_USERNAME=username
+# SPARK_PROXY_PASSWORD=password
+
+# =============================================================================
+# LOGGING CONFIGURATION
+# =============================================================================
+
+# Logger type (default: console, use json for Docker/production)
+SPARK_LOGGER=console
+# SPARK_LOGGER=json  # Recommended for Docker/production
+
+# =============================================================================
+# EXAMPLE CONFIGURATIONS
+# =============================================================================
+
+# Development Setup (local testing)
+# SPARK_PROVIDER=openai
+# OPENAI_API_KEY=sk-your-key
+# SPARK_BROWSER=chrome
+# SPARK_HEADLESS=false
+# SPARK_DEBUG=true
+# SPARK_VISION=true
+
+# Production/Docker Setup (optimized performance)
+# SPARK_PROVIDER=openai
+# OPENAI_API_KEY=sk-your-key
+# SPARK_BROWSER=firefox
+# SPARK_HEADLESS=true
+# SPARK_BLOCK_ADS=true
+# SPARK_BLOCK_RESOURCES=image,media,font
+# SPARK_MAX_ITERATIONS=30
+# SPARK_LOGGER=json
+
+# OpenRouter with Vision (alternative AI provider)
+# SPARK_PROVIDER=openrouter
+# OPENROUTER_API_KEY=sk-or-your-key
+# SPARK_MODEL=anthropic/claude-3-sonnet
+# SPARK_VISION=true
+# SPARK_DEBUG=true
+
+# Corporate Proxy Setup
+# SPARK_PROXY=http://corporate-proxy:8080
+# SPARK_PROXY_USERNAME=corp-user
+# SPARK_PROXY_PASSWORD=corp-pass
+# SPARK_BROWSER=chrome
+# SPARK_HEADLESS=true
+
+# Remote Browser Setup (external browserless)
+# SPARK_PW_ENDPOINT=ws://browserless.company.com:3000
+# SPARK_BROWSER=chromium
+# SPARK_HEADLESS=true
+# SPARK_BLOCK_ADS=true
