+++ server/package.json
@@ -26,18 +26,18 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "@types/node": "^20.17.25",
-    "prettier": "^3.5.3",
-    "tsx": "^4.20.2",
-    "typescript": "^5.8.2",
+    "@types/node": "^20.19.4",
+    "prettier": "^3.6.2",
+    "tsx": "^4.20.3",
+    "typescript": "^5.8.3",
     "vitest": "^1.6.1"
   },
   "dependencies": {
     "@ai-sdk/openai": "^1.3.22",
-    "@hono/node-server": "^1.14.4",
+    "@hono/node-server": "^1.15.0",
     "ai": "^4.3.16",
-    "dotenv": "^16.5.0",
-    "hono": "^4.7.11",
+    "dotenv": "^16.6.1",
+    "hono": "^4.8.3",
     "spark": "file:.."
   },
   "packageManager": "pnpm@9.0.0"
