+++ server/src/config.ts
@@ -1,0 +1,2 @@
+// Re-export shared config system for server
+export { config, ConfigManager, type SparkConfig } from "spark";
