+++ server/src/provider.ts
@@ -1,0 +1,2 @@
+// Re-export shared provider system for server
+export { createAIProvider, getAIProviderInfo } from "spark";
